implement logic to extract tagged users
