+++ lib/blocs/chats/messages/message_cubit.dart
@@ -4,7 +4,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../model/chat/message_model.dart';
 import '../../../services/chat/message_service.dart';
-import '../../../utils/app_utils.dart';
+import '../../../utils/chat_utils.dart';
 import 'message_state.dart';
 
 class MessageCubit extends Cubit<MessageState> {
