+++ lib/utils/app_utils.dart
@@ -1,22 +1,12 @@
-import 'dart:io';
-
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_link_previewer/flutter_link_previewer.dart';
-import 'package:intl/intl.dart';
-import 'package:linkify/linkify.dart';
-import 'package:pure/utils/global_utils.dart';
-import 'package:pure/views/widgets/snackbars.dart';
 import 'package:timeago/timeago.dart' as timeago;
-import 'package:flutter_chat_types/flutter_chat_types.dart' show PreviewData;
 import 'package:url_launcher/url_launcher.dart';
 import 'package:uuid/uuid.dart';
 
-import '../model/chat/chat_model.dart';
-import '../model/chat/message_model.dart';
 import '../model/pure_user_model.dart';
 import '../repositories/push_notification.dart';
 import '../services/user_service.dart';
+import '../views/widgets/snackbars.dart';
 
 // Generate a v1 (time-based) identifier
 String generateDatabaseId() {
@@ -42,53 +32,6 @@ String getFormattedTime(DateTime date) {
   return timeago.format(date, allowFromNow: true);
 }
 
-String groupDate(DateTime date) {
-  final filteredDate = DateTime(date.year, date.month, date.day);
-  if (DateTime.now().difference(filteredDate).inDays == 0) return "Today";
-  if (DateTime.now().difference(filteredDate).inDays == 1) return "Yesterday";
-  if (DateTime.now().difference(filteredDate).inDays < 7)
-    return DateFormat("EEEE").format(date);
-  if (DateTime.now().difference(filteredDate).inDays < 365)
-    return DateFormat.MMMEd().format(date);
-
-  return DateFormat.yMMMd().format(date);
-}
-
-String chatTime(final DateTime date) {
-  final filteredDate = DateTime(date.year, date.month, date.day);
-  if (DateTime.now().difference(filteredDate).inDays == 0)
-    return DateFormat.jm().format(date);
-  if (DateTime.now().difference(filteredDate).inDays == 1) return "Yesterday";
-  if (DateTime.now().difference(filteredDate).inDays < 7)
-    return DateFormat("EEEE").format(date);
-
-  return DateFormat.yMd().format(date);
-}
-
-// this method remove duplicate messages and still main order
-List<MessageModel> orderedSetForMessages(final List<MessageModel> messages) {
-  final result = messages.toList();
-  final messageIds = Set<String>();
-  result.retainWhere((x) => messageIds.add(x.messageId));
-  return result.toList();
-}
-
-// this method remove duplicate chats and still main order
-List<ChatModel> orderedSetForChats(final List<ChatModel> chats) {
-  final result = chats.toList();
-  final chatIds = Set<String>();
-  result.retainWhere((x) => chatIds.add(x.chatId));
-  return result.toList();
-}
-
-// this method remove duplicate Files and still main order
-List<File> orderedSetForFiles(final List<File> files) {
-  final result = files.toList();
-  final chatIds = Set<int>();
-  result.retainWhere((x) => chatIds.add(x.lengthSync()));
-  return result.toList();
-}
-
 // This function returns list of user connection
 
 List<String> getConnections(Map<String, ConnectionStatus> data) {
@@ -103,13 +46,6 @@ List<String> getConnections(Map<String, ConnectionStatus> data) {
   return connections.toList();
 }
 
-// gets lastDoc for messages
-DocumentSnapshot? getLastDoc(MessagesModel newMsg, MessagesModel oldMsg) {
-  return oldMsg.messages.length > newMsg.messages.length
-      ? oldMsg.lastDoc
-      : newMsg.lastDoc;
-}
-
 Future<void> launchIfCan(BuildContext context, String url) async {
   final result = await canLaunch(url);
   if (result)
@@ -118,21 +54,4 @@ Future<void> launchIfCan(BuildContext context, String url) async {
     final message = "Please install a browser that can open the link";
     showFailureFlash(context, message);
   }
-}
-
-Future<PreviewData?> getLinkPreviewData(String text) async {
-  final links = linkify(
-    text,
-    options: LinkifyOptions(humanize: false),
-    linkifiers: [UrlLinkifier()],
-  ).where((element) {
-    final link = element.text;
-    if (link.contains("http") || link.contains("https")) return true;
-    return false;
-  }).toList();
-
-  if (links.isNotEmpty)
-    return await getPreviewData(text).timeout(GlobalUtils.timeOutInDuration);
-  else
-    return null;
 }
