+++ lib/views/screens/chats/widget/one_to_one_card.dart
@@ -5,9 +5,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../../../blocs/bloc.dart';
 import '../../../../model/chat/chat_model.dart';
 import '../../../../model/pure_user_model.dart';
-import '../../../../utils/palette.dart';
-import '../../../../utils/app_utils.dart';
+import '../../../../utils/chat_utils.dart';
 import '../../../../utils/navigate.dart';
+import '../../../../utils/palette.dart';
 import '../../../widgets/avatar.dart';
 import '../../../widgets/user_profile_provider.dart';
 import '../messages/messages_screen.dart';
